she loves me but at what cost
