--- chadhmm/emissions/__init__.py
@@ -1,18 +1,0 @@
-
-__all__ = ['CategoricalEmissions', 
-           'GaussianEmissions',
-           'GammaEmissions',
-           'PoissonEmissions',
-           'MultinomialEmissions',
-           'BetaEmissions',
-           'ExponentialEmissions',
-           'GaussianMixtureEmissions']
-
-from .categorical import CategoricalEmissions
-from .gaussian import GaussianEmissions
-from .gamma import GammaEmissions
-from .poisson import PoissonEmissions
-from .multinomial import MultinomialEmissions
-from .gaussian_mix import GaussianMixtureEmissions
-from .beta import BetaEmissions
-from .exponential import ExponentialEmissions
