--- chadhmm/emissions/beta.py
@@ -1,6 +1,0 @@
-import torch
-from torch.distributions import Beta
-from .base_emiss import BaseEmission
-
-
-class BetaEmissions(BaseEmission): ... # TODO: we need to implement this
