--- chadhmm/emissions/exponential.py
@@ -1,6 +1,0 @@
-import torch
-from torch.distributions import Exponential
-from .base_emiss import BaseEmission
-
-
-class ExponentialEmissions(BaseEmission): ... # TODO: we need to implement this
