--- chadhmm/emissions/gamma.py
@@ -1,5 +1,0 @@
-import torch
-from torch.distributions import Gamma
-from .base_emiss import BaseEmission
-
-class GammaEmissions(BaseEmission): ... # TODO: we need to implement this
