--- chadhmm/emissions/multinomial.py
@@ -1,6 +1,0 @@
-import torch
-from torch.distributions import Multinomial
-from .base_emiss import BaseEmission
-
-
-class MultinomialEmissions(BaseEmission): ... # TODO: we need to implement this
