+++ requirements.txt
@@ -2,4 +2,3 @@ torch==2.0.1
 numpy==1.22.1
 scikit-learn==1.3.0
 matplotlib==3.7.2
-prettytable==3.8.0
